+++ CHANGELOG.md
@@ -37,3 +37,4 @@ Changes to the lab will be detailed in this log file.
 - Implemented routing on player-portfolio page
 - Set temporary default player-portfolio view to be blank
 - Implemented stock dropdown generation
+- Formatted stock history and player portfolio tables